integration: rename enforce_coordinator.rs -> coordinator.rs
The file will soon contain more coordinator-related tests than only
enforcing-related ones.
